// src/app/models/reservation.model.ts
class ReservationModel {
  ruc: string;
  socialname: string;
  rubro: string;
  webpage: string;
  address: string;
  city: string;
  country: string;
  legal: string;
  exhibitor: string;
  charge: string;
  phone: string;
  anexo: string;
  email: string;
  standname: string;
  detail: string;
  number: string;
  numberplan: string;
 
  constructor() {
    
  }
}

export default ReservationModel;