// src/app/models/reservationview.model.ts
class ReservationView {
  EVENT: {
    urlbackgroundimg: string;
    reservationtitle: string;
    personalinformation: string;
    registerpaymenttxt: string;
    registerreservationurl: string;
    contacttitle: string;
    contactphone: string;
    contact: string;
    contactbuttontxt: string;
    contactbuttonurl: string;
    contactplace: string;
    placeholder: {
      strName: string;
      strLastname: string;
      strMail: string;
      strMessage: string;
    },
    connectors: {
      strof: string;
      strto: string;
    },
    latitude: string;
    longitude: string;
    socialnetworks: {
      strlinkedinurl: string;
      strlinkedinlogo: string;
      strfacebookurl: string;
      strfacebooklogo: string;
      stryoutubeurl: string;
      stryoutubelogo: string;
      strtwitterurl: string;
      strtwitterlogo: string;
    }
  }
}



export default ReservationView;